fix(openOrders): map all out-of-range limit values to -1102, not -1130
Parse limit parameter as i64 before validating range, then cast to u16
after the range check passes. This ensures values like limit=65536
(above u16::MAX) return the spec-required HTTP 400 / code -1102,
not the previous -1130 parameter parsing error.

- Update GetOpenOrdersUseCase::execute to accept Option<i64> for limit
- Adjust range validation to compare against i64::from(OPEN_ORDERS_MAX_LIMIT)
- Update HTTP handler to parse limit as i64
- Add regression test for limit=65536 returning -1102

+++ crates/application/src/lib.rs
@@ -263,7 +263,7 @@ where
         ctx: &AuthContext,
         market_address: Option<MarketAddress>,
         symbol: Option<Symbol>,
-        limit: Option<u16>,
+        limit: Option<i64>,
         cursor: Option<&str>,
     ) -> Result<OpenOrdersPage, anyhow::Error> {
         let market = match (market_address, symbol) {
@@ -276,7 +276,7 @@ where
 
         let limit = match limit {
             None => OPEN_ORDERS_DEFAULT_LIMIT,
-            Some(v) if (1..=OPEN_ORDERS_MAX_LIMIT).contains(&v) => v,
+            Some(v) if (1..=i64::from(OPEN_ORDERS_MAX_LIMIT)).contains(&v) => v as u16,
             Some(_) => return Err(anyhow::anyhow!(DomainError::MissingParameter)),
         };
